git-svn-id: svn://192.168.1.10/public/CoffeeMud@20 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/interfaces/MOB.java
@@ -16,8 +16,8 @@ public interface MOB
 	public String password();
 	/** update USER information */
 	public void setUserInfo(String newUsername, String newPassword, Calendar newCalendar);
-	public void setChannelMask(long newMask);
-	public long getChannelMask();
+	public void setChannelMask(int newMask);
+	public int getChannelMask();
 	public boolean readSysopMsgs();
 	public void toggleReadSysopMsgs();
